requirements.txt
python-dotenv==0.21.0

py-cord==2.2.2
canvasapi==3.0.0

requests==2.28.1
PyPDF2==2.11.1
pytz==2022.4

openai==0.23.1
Pillow==9.2.0
torch==1.12.0
torchvision==0.13.0
transformers==4.23.1